backend/requirements.txt
web3==7.4.0
moralis==0.1.49
pandas
tqdm
python-dotenv
beautifulsoup4